Altered entity fetching approach to allow more simple, declarative reducers.
