+++ src/actionTypes/albumActionTypes.js
@@ -3,6 +3,4 @@ export const FETCH_ALBUM_SUCCESS = 'FETCH_ALBUM_SUCCESS';
 export const FETCH_ALBUM_FAILED = 'FETCH_ALBUM_FAILED';
 export const FETCH_ALBUM_ABORT = 'FETCH_ALBUM_ABORT';
 
-export const STORE_ALBUM = 'STORE_ALBUM';
-
 export const STORE_ALBUM_TRACK_IDS = 'STORE_ALBUM_TRACK_IDS';
