+++ src/actionTypes/artistActionTypes.js
@@ -3,11 +3,6 @@ export const FETCH_ARTIST_SUCCESS = 'FETCH_ARTIST_SUCCESS';
 export const FETCH_ARTIST_FAILED = 'FETCH_ARTIST_FAILED';
 export const FETCH_ARTIST_ABORT = 'FETCH_ARTIST_ABORT';
 
-export const STORE_ARTISTS_PROFILE = 'STORE_ARTISTS_PROFILE';
-export const STORE_ARTISTS_TOP_TRACKS = 'STORE_ARTISTS_TOP_TRACKS';
-export const STORE_ARTISTS_RELATED_ARTISTS = 'STORE_ARTISTS_RELATED_ARTISTS';
-export const STORE_ARTISTS_ALBUMS = 'STORE_ARTISTS_ALBUMS';
-
 export const STORE_ARTIST_TOP_TRACK_IDS = 'STORE_ARTIST_TOP_TRACK_IDS';
 export const STORE_ARTIST_ALBUM_IDS = 'STORE_ARTIST_ALBUM_IDS';
 export const STORE_ARTIST_RELATED_ARTIST_IDS = 'STORE_ARTIST_RELATED_ARTIST_IDS';
