+++ src/actionTypes/categoryActionTypes.js
@@ -3,5 +3,4 @@ export const FETCH_CATEGORY_SUCCESS = 'FETCH_CATEGORY_SUCCESS';
 export const FETCH_CATEGORY_FAILED = 'FETCH_CATEGORY_FAILED';
 export const FETCH_CATEGORY_ABORT = 'FETCH_CATEGORY_ABORT';
 
-export const STORE_CATEGORY_INFO = 'STORE_CATEGORY_INFO';
-export const STORE_CATEGORIES_PLAYLISTS = 'STORE_CATEGORIES_PLAYLISTS';
+export const STORE_CATEGORY_PLAYLIST_IDS = 'STORE_CATEGORY_PLAYLIST_IDS';
