+++ src/actionTypes/highlightsActionTypes.js
@@ -3,8 +3,4 @@ export const FETCH_HIGHLIGHTS_SUCCESS = 'FETCH_HIGHLIGHTS_SUCCESS';
 export const FETCH_HIGHLIGHTS_FAILED = 'FETCH_HIGHLIGHTS_FAILED';
 export const FETCH_HIGHLIGHTS_ABORT = 'FETCH_HIGHLIGHTS_ABORT';
 
-export const STORE_NEW_RELEASES = 'STORE_NEW_RELEASES';
-export const STORE_FEATURED_PLAYLISTS = 'STORE_FEATURED_PLAYLISTS';
-export const STORE_CATEGORIES = 'STORE_CATEGORIES';
-
 export const STORE_HIGHLIGHTS = 'STORE_HIGHLIGHTS';
