+++ src/actionTypes/playerActionTypes.js
@@ -1,27 +1,31 @@
-export const SELECT_TRACK_REQUEST = 'SELECT_TRACK_REQUEST';
-export const SELECT_TRACK_SUCCESS = 'SELECT_TRACK_SUCCESS';
-export const SELECT_TRACK_FAILED = 'SELECT_TRACK_FAILED';
+// Standard player related actions
+
+export const STANDARD_SELECT_TRACK = 'STANDARD_SELECT_TRACK';
+export const STANDARD_RESUME_PLAYER = 'STANDARD_RESUME_PLAYER';
+export const STANDARD_PAUSE_PLAYER = 'STANDARD_PAUSE_PLAYER';
+export const STANDARD_SKIP_FORWARDS = 'STANDARD_SKIP_FORWARDS';
+export const STANDARD_SKIP_BACKWARDS = 'STANDARD_SKIP_BACKWARDS';
+export const STANDARD_SET_SHUFFLE = 'STANDARD_SET_SHUFFLE';
+export const STANDARD_SET_REPEAT = 'STANDARD_SET_REPEAT';
 
 
-export const PLAY_CURRENT_TRACK = 'PLAY_CURRENT_TRACK';
-export const PAUSE_CURRENT_TRACK = 'PAUSE_CURRENT_TRACK';
-export const SKIP_TO_NEXT_TRACK = 'SKIP_TO_NEXT_TRACK';
-export const SKIP_TO_PREVIOUS_TRACK = 'SKIP_TO_PREVIOUS_TRACK';
-export const SEEK_POSITION_IN_TRACK = 'SEEK_POSITION_IN_TRACK';
 
-export const SET_SHUFFLE_REQUEST = 'SET_SHUFFLE_REQUEST';
-export const SET_SHUFFLE_SUCCESS = 'SET_SHUFFLE_SUCCESS';
-export const SET_SHUFFLE_FAILED = 'SET_SHUFFLE_FAILED';
-
-export const SET_REPEAT_REQUEST = 'SET_REPEAT_REQUEST';
-export const SET_REPEAT_SUCCESS = 'SET_REPEAT_SUCCESS';
-export const SET_REPEAT_FAILED = 'SET_REPEAT_FAILED';
-
+// SDK player related actions
 export const CONFIRM_SDK_AVAILABLE = 'CONFIRM_SDK_AVAILABLE';
 
-export const UPDATE_PLAYER_STATE = 'UPDATE_PLAYER_STATE';
-
 export const SDK_UPDATE_PLAYER_STATE = 'SDK_UPDATE_PLAYER_STATE';
+
+// The rest of these SDK actions currently aren't being used within the reducer, since all of the player
+// updates for the SDK player are driven by the SDK_UPDATE_PLAYER_STATE action. The below actions can still
+// be useful for debugging however.
+
+export const SDK_SET_SHUFFLE_REQUEST = 'SDK_SET_SHUFFLE_REQUEST';
+export const SDK_SET_SHUFFLE_SUCCESS = 'SDK_SET_SHUFFLE_SUCCESS';
+export const SDK_SET_SHUFFLE_FAILED = 'SDK_SET_SHUFFLE_FAILED';
+
+export const SDK_SET_REPEAT_REQUEST = 'SDK_SET_REPEAT_REQUEST';
+export const SDK_SET_REPEAT_SUCCESS = 'SDK_SET_REPEAT_SUCCESS';
+export const SDK_SET_REPEAT_FAILED = 'SDK_SET_REPEAT_FAILED';
 
 export const SDK_SELECT_TRACK_SUCCESS = 'SDK_SELECT_TRACK_SUCCESS';
 export const SDK_SELECT_TRACK_FAILED = 'SDK_SELECT_TRACK_FAILED';
@@ -37,19 +41,3 @@ export const SDK_SKIP_FORWARDS_FAILED = 'SDK_SKIP_FORWARDS_FAILED';
 
 export const SDK_SKIP_BACKWARDS_SUCCESS = 'SDK_SKIP_BACKWARDS_SUCCESS';
 export const SDK_SKIP_BACKWARDS_FAILED = 'SDK_SKIP_BACKWARDS_FAILED';
-
-export const SDK_SET_SHUFFLE_REQUEST = 'SDK_SET_SHUFFLE_REQUEST';
-export const SDK_SET_SHUFFLE_SUCCESS = 'SDK_SET_SHUFFLE_SUCCESS';
-export const SDK_SET_SHUFFLE_FAILED = 'SDK_SET_SHUFFLE_FAILED';
-
-export const SDK_SET_REPEAT_REQUEST = 'SDK_SET_REPEAT_REQUEST';
-export const SDK_SET_REPEAT_SUCCESS = 'SDK_SET_REPEAT_SUCCESS';
-export const SDK_SET_REPEAT_FAILED = 'SDK_SET_REPEAT_FAILED';
-
-export const STANDARD_SELECT_TRACK = 'STANDARD_SELECT_TRACK';
-export const STANDARD_RESUME_PLAYER = 'STANDARD_RESUME_PLAYER';
-export const STANDARD_PAUSE_PLAYER = 'STANDARD_PAUSE_PLAYER';
-export const STANDARD_SKIP_FORWARDS = 'STANDARD_SKIP_FORWARDS';
-export const STANDARD_SKIP_BACKWARDS = 'STANDARD_SKIP_BACKWARDS';
-export const STANDARD_SET_SHUFFLE = 'STANDARD_SET_SHUFFLE';
-export const STANDARD_SET_REPEAT = 'STANDARD_SET_REPEAT';
