+++ src/actionTypes/playlistActionTypes.js
@@ -3,9 +3,6 @@ export const FETCH_PLAYLIST_SUCCESS = 'FETCH_PLAYLIST_SUCCESS';
 export const FETCH_PLAYLIST_FAILED = 'FETCH_PLAYLIST_FAILED';
 export const FETCH_PLAYLIST_ABORT = 'FETCH_PLAYLIST_ABORT';
 
-export const STORE_PLAYLIST = 'STORE_PLAYLIST';
-export const STORE_PLAYLIST_FOLLOW_STATUS = 'STORE_PLAYLIST_FOLLOW_STATUS';
-
 export const UPDATE_PLAYLIST_NAME_REQUEST = 'UPDATE_PLAYLIST_NAME_REQUEST';
 export const UPDATE_PLAYLIST_NAME_SUCCESS = 'UPDATE_PLAYLIST_NAME_SUCCESS';
 export const UPDATE_PLAYLIST_NAME_FAILED = 'UPDATE_PLAYLIST_NAME_FAILED';
@@ -25,3 +22,6 @@ export const REMOVE_TRACK_FROM_PLAYLIST_FAILED = 'REMOVE_TRACK_FROM_PLAYLIST_FAI
 export const CREATE_PLAYLIST_REQUEST = 'CREATE_PLAYLIST_REQUEST';
 export const CREATE_PLAYLIST_SUCCESS = 'CREATE_PLAYLIST_SUCCESS';
 export const CREATE_PLAYLIST_FAILED = 'CREAT_PLAYLIST_FAILED';
+
+export const STORE_USER_FOLLOWING_PLAYLIST = 'STORE_USER_FOLLOWING_PLAYLIST';
+export const STORE_PLAYLIST_TRACK_IDS = 'STORE_PLAYLIST_TRACK_IDS';
