+++ src/actionTypes/userActionTypes.js
@@ -4,9 +4,3 @@ export const FETCH_USER_FAILED = 'FETCH_USER_FAILED';
 export const FETCH_USER_ABORT = 'FETCH_USER_ABORT';
 
 export const STORE_USERS_PROFILE = 'STORE_USERS_PROFILE';
-export const STORE_USERS_TOP_ARTISTS = 'STORE_USERS_TOP_ARTISTS';
-export const STORE_USERS_RECENT_TRACKS = 'STORE_USERS_RECENT_TRACKS';
-export const STORE_USERS_PLAYLISTS = 'STORE_USERS_PLAYLISTS';
-export const STORE_USERS_FOLLOWED_ARTISTS = 'STORE_USERS_FOLLOWED_ARTISTS';
-
-export const SET_MARKET = 'SET_MARKET';
