+++ src/actions/highlightsActions.js
@@ -1,7 +1,7 @@
 import * as actionTypes from '../actionTypes';
+import { storeAlbums, storeArtists, storePlaylists, storeCategories } from './entityActions';
 import { normalize, schema } from 'normalizr';
 import axios from 'axios';
-import { getUsersMarket } from './userActions';
 
 const fetchHighlightsRequest = () => ({
     type: actionTypes.FETCH_HIGHLIGHTS_REQUEST
@@ -25,30 +25,25 @@ const fetchHighlightsAbort = () => ({
     type: actionTypes.FETCH_HIGHLIGHTS_ABORT
 });
 
-const storeNewReleases = (albumObjects, albumIds, artistObjects) => ({
-    type: actionTypes.STORE_NEW_RELEASES,
+const storeHighlights = (newReleaseIds, featuredPlaylistIds, categoryIds) => ({
+    type: actionTypes.STORE_HIGHLIGHTS,
     payload: {
-        albumObjects,
-        albumIds,
-        artistObjects
-    }
-});
-
-const storeFeaturedPlaylists = (playlistObjects, playlistIds) => ({
-    type: actionTypes.STORE_FEATURED_PLAYLISTS,
-    payload: {
-        playlistObjects,
-        playlistIds
-    }
-});
-
-const storeCategories = (categoryObjects, categoryIds) => ({
-    type: actionTypes.STORE_CATEGORIES,
-    payload: {
-        categoryObjects,
+        newReleaseIds,
+        featuredPlaylistIds,
         categoryIds
     }
 });
+
+const artistSchema = new schema.Entity('artists');
+const albumSchema = new schema.Entity('albums', { artists: [artistSchema] });
+const playlistSchema = new schema.Entity('playlists');
+const categorySchema = new schema.Entity('categories');
+
+
+// const normalizedData = normalize(response.data.albums.items, [albumSchema]);
+// const normalizedData = normalize(response.data.playlists.items, [playlistSchema]);
+// const normalizedData = normalize(response.data.categories.items, [categorySchema]);
+
 
 const fetchNewReleases = (token, market) => async (dispatch) => {
     try {
@@ -58,14 +53,7 @@ const fetchNewReleases = (token, market) => async (dispatch) => {
                 'Authorization': `Bearer ${token}`
             }
         });
-        const artistSchema = new schema.Entity('artists');
-        const albumSchema = new schema.Entity('albums', { artists: [artistSchema] });
-        const normalizedData = normalize(response.data.albums.items, [albumSchema]);
-        dispatch(storeNewReleases(
-            normalizedData.entities.albums,
-            normalizedData.result,
-            normalizedData.entities.artists
-        ));
+        return normalize(response.data.albums.items, [albumSchema]); 
     } catch (err) {
         throw new Error(err);
     }
@@ -79,12 +67,7 @@ const fetchFeaturedPlaylists = (token, market) => async (dispatch) => {
                 'Authorization': `Bearer ${token}`
             }
         });
-        const playlistSchema = new schema.Entity('playlists');
-        const normalizedData = normalize(response.data.playlists.items, [playlistSchema]);
-        dispatch(storeFeaturedPlaylists(
-            normalizedData.entities.playlists,
-            normalizedData.result
-        ));
+        return normalize(response.data.playlists.items, [playlistSchema]);
     } catch (err) {
         throw new Error(err);
     }
@@ -98,34 +81,78 @@ const fetchCategories = (token, market) => async (dispatch) => {
                 'Authorization': `Bearer ${token}`
             }
         });
-        const categorySchema = new schema.Entity('categories');
-        const normalizedData = normalize(response.data.categories.items, [categorySchema]);
-        dispatch(storeCategories(
-            normalizedData.entities.categories,
-            normalizedData.result
-        ));
+        return normalize(response.data.categories.items, [categorySchema]);
     } catch (err) {
         throw new Error(err);
     }
 }
 
+const destructureData = (resolvedPromiseArr) => {
+    const [
+        {
+            entities: {
+                albums: albumEntities,
+                artists: artistEntities
+            },
+            result: newReleaseIds
+        },
+        {
+            entities: {
+                playlists: playlistEntities
+            },
+            result: featuredPlaylistIds
+        },
+        {
+            entities: {
+                categories: categoryEntities
+            },
+            result: categoryIds
+        }
+    ] = resolvedPromiseArr;
+
+    return {
+        artistEntities,
+        albumEntities,
+        playlistEntities,
+        categoryEntities,
+        newReleaseIds,
+        featuredPlaylistIds,
+        categoryIds
+    }
+}
+
 export const fetchHighlights = () => async (dispatch, getState) => {
     const token = getState().accessToken.token;
-    const market = getState().user.country || await dispatch(getUsersMarket(token));
+    const market = getState().user.country;
     const highlights = getState().highlights;
     if (highlights.fullHighlightsFetched && Date.now() - highlights.lastFetchedAt <= 3600000) {
         return dispatch(fetchHighlightsAbort());
     }
     dispatch(fetchHighlightsRequest());
-    Promise.all([
-        dispatch(fetchNewReleases(token, market)),
-        dispatch(fetchFeaturedPlaylists(token, market)),
-        dispatch(fetchCategories(token, market))
-    ])
-    .then(() => {
-        const timestamp = Date.now();
-        dispatch(fetchHighlightsSuccess(timestamp));
-    }, (err) => {
+
+    try {
+        const results = await Promise.all([
+            dispatch(fetchNewReleases(token, market)),
+            dispatch(fetchFeaturedPlaylists(token, market)),
+            dispatch(fetchCategories(token, market))
+        ]);
+        const {
+            albumEntities,
+            artistEntities,
+            playlistEntities,
+            categoryEntities,
+            newReleaseIds,
+            featuredPlaylistIds,
+            categoryIds
+        } = destructureData(results);
+        dispatch(storeAlbums(albumEntities));
+        dispatch(storeArtists(artistEntities));
+        dispatch(storePlaylists(playlistEntities));
+        dispatch(storeCategories(categoryEntities));
+        dispatch(storeHighlights(newReleaseIds, featuredPlaylistIds, categoryIds));
+        dispatch(fetchHighlightsSuccess(Date.now()));
+    } catch (err) {
         dispatch(fetchHighlightsFailed(err));
-    })
+    }
+
 }
