+++ src/actions/index.js
@@ -10,4 +10,4 @@ export * from './transitionActions';
 export * from './followActions';
 export * from './modalActions';
 export * from './playerActions';
-export * from './newArtistActions';
+
