+++ src/reducers/albumsReducer/entitiesReducer.js
@@ -6,23 +6,8 @@ const defaultState = {};
 const reducer = (state=defaultState, action) => {
     switch (action.type) {
 
-        case actionTypes.STORE_USERS_RECENT_TRACKS:
-            return addOrMerge(state, action.payload.albumObjects);
-
-        case actionTypes.STORE_ARTISTS_TOP_TRACKS:
-            return addOrMerge(state, action.payload.albumObjects);
-
-        case actionTypes.STORE_ARTISTS_ALBUMS:
-            return addOrMerge(state, action.payload.albumObjects);
-
-        case actionTypes.STORE_ALBUM:
-            return addOrMerge(state, action.payload.albumObject, action.payload.albumId);
-
-        case actionTypes.STORE_PLAYLIST:
-            return addOrMerge(state, action.payload.albumObjects);
-
-        case actionTypes.STORE_NEW_RELEASES:
-            return addOrMerge(state, action.payload.albumObjects);
+        case actionTypes.STORE_ALBUMS:
+            return addOrMerge(state, action.payload.albumEntities);
 
         default:
             return state;
