+++ src/reducers/albumsReducer/trackIdsReducer.js
@@ -4,11 +4,11 @@ const defaultState = {};
 
 const reducer = (state=defaultState, action) => {
     switch (action.type) {
-
-        case actionTypes.STORE_ALBUM:
+    
+        case actionTypes.STORE_ALBUM_TRACK_IDS:
             return {
                 ...state,
-                [action.payload.albumId]: action.payload.albumsTrackIds
+                [action.payload.ownerId]: action.payload.albumTrackIds
             };
     
         default:
