+++ src/reducers/categoriesReducer/entitiesReducer.js
@@ -7,10 +7,7 @@ const reducer = (state=defaultState, action) => {
     switch (action.type) {
 
         case actionTypes.STORE_CATEGORIES:
-            return addOrMerge(state, action.payload.categoryObjects);
-
-        case actionTypes.STORE_CATEGORY_INFO:
-            return addOrMerge(state, action.payload.categoryObject, action.payload.categoryId);
+            return addOrMerge(state, action.payload.categoryEntities);
 
         default:
             return state;
