+++ src/reducers/tracksReducer.js
@@ -6,20 +6,8 @@ const defaultState = {};
 const tracks = (state=defaultState, action) => {
     switch (action.type) {
 
-        case actionTypes.STORE_USERS_RECENT_TRACKS:
-            return {
-                ...state,
-                ...action.payload.trackObjects
-            }
-
-        case actionTypes.STORE_ARTISTS_TOP_TRACKS:
-            return addOrMerge(state, action.payload.trackObjects);
-
-        case actionTypes.STORE_ALBUM:
-            return addOrMerge(state, action.payload.trackObjects);
-
-        case actionTypes.STORE_PLAYLIST:
-            return addOrMerge(state, action.payload.trackObjects);
+        case actionTypes.STORE_TRACKS:
+            return addOrMerge(state, action.payload.trackEntities);
 
         default:
             return state;
