+++ src/reducers/userReducer.js
@@ -35,30 +35,6 @@ const user = (state=defaultState, action) => {
                 ...action.payload.usersProfile
             };
 
-        case actionTypes.STORE_USERS_TOP_ARTISTS:
-            return {
-                ...state,
-                topArtistsIds: [ ...action.payload.artistIds ]
-            };
-
-        case actionTypes.STORE_USERS_RECENT_TRACKS:
-            return {
-                ...state,
-                recentTracksIds: [ ...action.payload.trackIds ]
-            };
-
-        case actionTypes.STORE_USERS_PLAYLISTS:
-            return {
-                ...state,
-                playlistIds: [ ...action.payload.playlistIds ]
-            };
-
-        case actionTypes.STORE_USERS_FOLLOWED_ARTISTS:
-            return {
-                ...state,
-                followedArtistIds: [ ...action.payload.artistIds ]
-            };
-
         case actionTypes.FOLLOW_ARTIST_SUCCESS: 
             return {
                 ...state,
